still need to figure out why not re-rendering after adding task/moving tasks

+++ server/database/migrations/20200123115546_lists.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
       .references("boards.id")
       .notNullable();
     table.string("title");
-    table.decimal("position").notNullable();
+    table.integer("position").notNullable();
     table.timestamps(true, true);
   });
 };
